// src/app/model/siteStatus.ts
export interface ISitestatus{
    "sl.no": number;
    "siteName": string;
    "siteLabel": string;
    "street":string;
    "city":string;
    "district": number;
    "state": string;
    "industryType": string;
    "lastSynchronized": number;
    "siteStatus": string;
    "exceedance": number;
    "vendor": number;
    "siteId": string;
   
  }